Backups

+++ c3-platform/c3-backup/src/main/scala/org/aphreet/c3/platform/backup/impl/BackupManagerImpl.scala
@@ -30,22 +30,29 @@
 
 package org.aphreet.c3.platform.backup.impl
 
-import org.aphreet.c3.platform.backup.BackupManager
+import org.aphreet.c3.platform.backup._
 import org.aphreet.c3.platform.storage.{StorageIterator, Storage, StorageManager}
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Component
 import org.aphreet.c3.platform.config.{PropertyChangeEvent, SPlatformPropertyListener, PlatformConfigManager}
-import org.aphreet.c3.platform.task.{IterableTask, TaskManager, Task}
+import org.aphreet.c3.platform.task.{TaskScheduleObserver, IterableTask, TaskManager, Task}
 import org.aphreet.c3.platform.resource.{ResourceAddress, Resource}
 import org.aphreet.c3.platform.common.Path
-import org.aphreet.c3.platform.access.{ResourceAddedMsg, AccessMediator}
+import org.aphreet.c3.platform.access.AccessMediator
 import org.aphreet.c3.platform.filesystem.FSManager
 import java.io.File
-import java.util
-import collection.mutable.ListBuffer
+import scala.collection.mutable.ListBuffer
+import javax.annotation.{PreDestroy, PostConstruct}
+import org.apache.commons.logging.LogFactory
+import scala._
+import ssh.SftpConnector
+import scala.Some
+import org.aphreet.c3.platform.access.ResourceAddedMsg
+import org.aphreet.c3.platform.backup.BackupLocation
+import org.apache.commons.codec.binary.StringUtils
 
 @Component("backupManager")
-class BackupManagerImpl extends BackupManager with SPlatformPropertyListener{
+class BackupManagerImpl extends BackupManager with SPlatformPropertyListener with TaskScheduleObserver {
 
   val BACKUP_LOCATION = "c3.platform.backup.location"
 
@@ -64,21 +71,37 @@ class BackupManagerImpl extends BackupManager with SPlatformPropertyListener{
   @Autowired
   var taskManager:TaskManager = _
 
-  def createBackup(){
-    val backupDirectory = configManager.getPlatformProperties.get(BACKUP_LOCATION) match {
-      case Some(value) => value
-      case None => throw new IllegalStateException("Can't create backup without " + BACKUP_LOCATION + " property")
-    }
+  @Autowired
+  var configAccessor: BackupConfigAccessor = _
 
-    val storages = storageManager.listStorages
+  var targets : List[BackupLocation] = null
 
-    val task = new BackupTask(storages, filesystemManager, new Path(backupDirectory))
+  val log = LogFactory getLog getClass
+
+
+  @PostConstruct
+  def init(){
+    targets = configAccessor.load
+
+    targets.foreach(target => target.schedule.foreach(s => scheduleBackup(target, s)))
+  }
+
+  def createBackup(targetId : String) {
+    val target = getBackupLocation(targetId)
+
+    val task = new BackupTask(storageManager, filesystemManager, target)
 
     taskManager.submitTask(task)
   }
 
-  def restoreBackup(location:String){
-    val backup = Backup.open(Path(location))
+  def restoreBackup(targetId: String, name: String) {
+    val target = getBackupLocation(targetId)
+
+    val backup : AbstractBackup = target.backupType match {
+      case "local" => Backup.open(Path(target.folder + "/" + name))
+      case "remote" => RemoteBackup.open(name, target)
+      case _ => throw new IllegalStateException("Wrong target type")
+    }
 
     storageManager.resetStorages()
 
@@ -86,19 +109,152 @@ class BackupManagerImpl extends BackupManager with SPlatformPropertyListener{
     taskManager.submitTask(task)
   }
 
-  def listBackups(folderPath:String) : List[String] = {
-    val folder = new File(folderPath)
+  def scheduleBackup(targetId:String, crontabSchedule: String) {
+    val target = getBackupLocation(targetId)
+
+    scheduleBackup(target, crontabSchedule)
+
+    target.schedule ::= crontabSchedule
+    configAccessor.update(l => targets)
+  }
+
+  private def scheduleBackup(target: BackupLocation, crontabSchedule: String){
+    val task = new BackupTask(storageManager, filesystemManager, target)
+
+    taskManager.scheduleTask(task, crontabSchedule)
+
+    task.addObserver(this)
+  }
+
+  def listBackups(targetId:String) : List[String] = {
+    val target = getBackupLocation(targetId)
+
+    target.backupType match {
+      case "local" => listLocalBackups(target)
+      case "remote" => listRemoteBackups(target)
+      case _ => throw new IllegalStateException("Wrong target type")
+    }
+  }
+
+  def listLocalBackups(target: BackupLocation) : List[String] = {
     val listBuffer = new ListBuffer[String]()
+    val folder = new File(target.folder)
 
     if (folder.exists() && folder.isDirectory) {
       val filesList = folder.listFiles()
 
       filesList
         .filter( file => file.isFile && file.getName.endsWith(".zip") && Backup.hasValidChecksum(file.getAbsolutePath))
-        .foreach(file => listBuffer += file.getAbsolutePath)
+        .foreach( file => listBuffer += file.getAbsolutePath)
     }
 
     listBuffer.toList
+  }
+
+  def listRemoteBackups(target: BackupLocation) : List[String] = {
+    val listBuffer = new ListBuffer[String]()
+
+    val connector = new SftpConnector(target.host, target.user, target.privateKey)
+    connector.connect()
+
+    val allFilesNames = connector.listFiles(target.folder)
+
+    log.info("All file Names in folder:")
+    for (name <- allFilesNames) {
+      log.info(name)
+    }
+
+    allFilesNames
+      .filter( fileName => fileName.endsWith(".zip")
+                             && !allFilesNames.find(str => str.equals(fileName + ".md5")).isEmpty
+                             && RemoteBackup.hasValidChecksum(connector, target.folder, fileName))
+
+      .foreach( fileName => listBuffer += fileName)
+
+    connector.disconnect()
+
+    log.info("Filtered file names in folder:")
+    for (name <- listBuffer) {
+      log.info(name)
+    }
+
+    listBuffer.toList
+  }
+
+
+  def createLocalTarget(id: String, path: String) {
+    if (existsTargetId(id)) {
+      throw new IllegalArgumentException("There is already a target with specified ID")
+    }
+
+    val target = LocalBackupLocation.create(id, path)
+    targets ::= target
+    configAccessor.update(l => targets)
+  }
+
+  def createRemoteTarget(id: String, host: String, user: String, path: String, privateKeyFile: String) {
+    if (existsTargetId(id)) {
+      throw new IllegalArgumentException("There is already a target with specified ID")
+    }
+
+    val target = RemoteBackupLocation.create(id, host, user, path, privateKeyFile)
+    targets ::= target
+    configAccessor.update(l => targets)
+  }
+
+  def removeTarget(id: String) {
+    val target = getBackupLocation(id)
+    targets = targets.diff(List(target))
+    configAccessor.update(l => targets)
+  }
+
+  def listTargets() : List[BackupLocation] = targets
+
+  def showTargetInfo(targetId : String) : BackupLocation = getBackupLocation(targetId)
+
+  def getBackupLocation(targetId : String) : BackupLocation = {
+    val maybeLocation = targets.find(target => target.id.equals(targetId))
+
+    val backupLocation = maybeLocation match {
+      case Some(value) => value
+      case None => {
+        if (targetId.forall(_.isDigit)) {
+          val num = targetId.toInt
+          if (num >= 1 && num <= targets.size) {
+            targets(num - 1)
+          } else {
+            throw new IllegalArgumentException("Target number is too big or less than 1")
+          }
+        } else {
+          throw new IllegalArgumentException("Argument is not a target id nor target number")
+        }
+      }
+    }
+
+    backupLocation
+  }
+
+  def updateSchedule(task: Task, newSchedule: String) {
+    if (task.isInstanceOf[BackupTask]) {
+      val backupTask = task.asInstanceOf[BackupTask]
+      val target = getBackupLocation(backupTask.target.id)
+      val oldSchedule = task.getSchedule
+
+      target.schedule = target.schedule.diff(List(oldSchedule))
+      if (newSchedule != null && !newSchedule.equals("")) {
+        target.schedule ::= newSchedule
+      }
+
+      configAccessor.update(l => targets)
+    }
+  }
+
+  def existsTargetId(targetId : String) : Boolean = {
+    targets.find(target => target.id.equals(targetId)) match {
+
+      case Some(value) => true
+      case None => false
+    }
   }
 
   def propertyChanged(event: PropertyChangeEvent) {}
@@ -106,15 +262,18 @@ class BackupManagerImpl extends BackupManager with SPlatformPropertyListener{
   def defaultValues = Map(BACKUP_LOCATION -> System.getProperty("user.home"))
 }
 
-class BackupTask(val storages:List[Storage], val fsManager:FSManager, val directory:Path) extends Task{
+class BackupTask(val storageManager: StorageManager, val fsManager:FSManager, val target: BackupLocation) extends Task{
 
   var iterator:StorageIterator = null
 
-  var storagesToProcess = storages
+  var storagesToProcess: List[Storage] = null
 
-  var backup:Backup = null
+  var backup : AbstractBackup = null
 
   var backupName:Path = null
+
+  var isLocal : Boolean = true
+
 
   protected override def step() {
     if (iterator == null){
@@ -148,22 +307,41 @@ class BackupTask(val storages:List[Storage], val fsManager:FSManager, val direct
       backup.close()
     }
 
-    if (backupName.file.exists()){
-      backupName.file.delete()
+    if (isLocal) {
+      if (backupName.file.exists()){
+        backupName.file.delete()
+      }
     }
   }
 
   override def preStart(){
+    storagesToProcess = storageManager.listStorages
 
-    if(!directory.file.exists()){
-      directory.file.mkdirs()
+    isLocal = target.backupType match {
+      case "local" => true
+      case "remote" => false
+      case _ => throw new IllegalStateException("Wrong type of target")
     }
 
-    backupName = directory.append("backup-" + System.currentTimeMillis() + ".zip")
+    if (isLocal) {
+      val directory = new Path(target.folder)
+      if(!directory.file.exists()){
+        directory.file.mkdirs()
+      }
 
-    log.info("Creating backup file " + backupName.stringValue)
+      backupName = directory.append("backup-" + System.currentTimeMillis() + ".zip")
 
-    backup = Backup.create(backupName)
+      log.info("Creating backup file " + backupName.stringValue)
+
+      backup = Backup.create(backupName)
+
+    } else {
+      backupName = new Path("backup-" + System.currentTimeMillis() + ".zip")
+
+      log.info("Creating backup file " + backupName.stringValue)
+
+      backup = RemoteBackup.create(backupName.stringValue, target)
+    }
   }
 
   override def postComplete(){
@@ -183,29 +361,29 @@ class BackupTask(val storages:List[Storage], val fsManager:FSManager, val direct
   }
 }
 
-class RestoreTask(val storageManager:StorageManager, val accessMediator:AccessMediator, val fsManager:FSManager, val backup:Backup)
+
+class RestoreTask(val storageManager: StorageManager, val accessMediator: AccessMediator,
+                  val fsManager: FSManager, val backup: AbstractBackup)
   extends IterableTask[Resource](backup){
 
   override def processElement(resource:Resource){
 
-    if(log.isDebugEnabled)
+    if (log.isDebugEnabled)
       log.debug("Importing resource " + resource.address)
 
     storageManager.storageForAddress(ResourceAddress(resource.address)).update(resource)
     accessMediator ! ResourceAddedMsg(resource, 'BackupManager)
   }
 
-  override
-  protected def preStart(){
+  override protected def preStart(){
     val fsRoots = backup.readFileSystemRoots
 
     for ((domain, address) <- fsRoots){
-      try{
+      try {
         fsManager.importFileSystemRoot(domain, address)
-      }catch{
+      } catch {
         case e:Throwable => log.warn("Failed to import file system root for " + domain + ": " + address, e)
       }
     }
-
   }
 }
